+++ sheep/tools/buildingManager.cpp
@@ -1,0 +1,10 @@
+#include "buildingManager.h"
+
+namespace sheep
+{
+    // enum BuildingManager::BuildingType
+    // {
+    //     standard = 0,
+    //     polygon = 1
+    // };
+}