+++ src/cityscape.h
@@ -4,7 +4,10 @@
 #include "../sheep/sheep.h"
 #include "cube.h"
 #include "texturedCube.h"
-#include "buildings/normalBuilding.h"
+#include "randomNumberGenerator.h"
+
+#include <iostream>
+#include <glm/gtx/string_cast.hpp>
 
 // Objects
 #include "cube.h"
@@ -20,12 +23,19 @@ class Cityscape: public Sample
         void update(float dt) override;
         void render(int width, int height) override;
 
+        // Regenerates the world
+        void regenerate();
+
     private:
         wolf::Program* mainProgram = 0;
         sheep::FirstPersonCamera* camera;
         GLfloat time = 0;
 
+        GLboolean isRegen = GL_FALSE;
+
         // TODO: Remove later
         std::vector<sheep::StandardBuilding*> standardBuildings;
+        std::vector<sheep::PolygonBuilding*> polygonBuildings;
+        TexturedCube* texturedCube1;
 
 };